fix: FINALLY! Found where the rotation problem was.
This was not a model problem or a rotation one. This was a more profound problem or matrix ordering and orientation which I fixed in other places so the display was somehow working. BUT when testing, I discovered that all coordinates where good so it must be display problem. So i change the camera controller to be working and broking the display. I changed the camera proj, etc. and finally tried to swap my quaternionToMatrix to the one wgpu proposed and it fixed the problem. All transform tests are now ok and verified + main.ts rotate successfully.